Basic wrapper over Secure Cell (sealed mode)
Here's how we can wrap Themis' C API in Rust. It's pretty
straightforward, isn't it? The only wacky point that I would
like to improve it error handling. Error codes are currently
hardcoded and are exported in raw form to the user. We should
wrap them and implement Error trait for that.