Secure Messages
This introduces wrappers for Secure Messages.

In order to make sign/verify mode to work we actually need to pass
null pointers into a function for messages. as_ptr() of slices returns
some non-null pointer so we had to tweak into_raw_parts() a bit. As
well as move it into common utility module.

Also, during testing I have found that Themis crashes (!) when it
encounters corrupted data. Well, that's a surprise. I am testing this
on macOS 10.12 with the current master version (82033534), but it seems
to crash with the released version 0.10 as well. We'll get back to that
later, I guess.