Move keygen to its own module
It does not make sense to keep it with Secure Messages
so move key generation to a separate module.

While we're here, introduce "newtypes" for private and
public keys so that the user can't mess them. Currently
we're still using byte slices in messages so we have a
temporary Deref implementation. We'll replace it with
AsRef together with interface update.