New Secure Cell interface
We replace raw function calls with proper Rusty interface.

The entry point is themis::secure_cell::SecureCell struct which
holds the master key and optional context. The types are generic
so you can use strings, byte slices, vectors, whatever--as long
as it is convertible into a byte slice via AsRef<[u8]>.

Note that the context is optional. There is one minor wacky point
with type inference when the context is not specified. We do not
care about specific context type in this case, but we need to tell
the compiler something so we use the same type as for the key.

After that SecureCell can be converted into one of the modes that
provide concrete encrypt/decrypt methods. The methods have different
signatures based on the expected results.

It is possible to trivially convert between Secure Cell types, but
this is not implemented at the moment as it does not seem to be that
useful. Usually one would create a specific type of secure cell and
use it with a specific master key and context.

See unit test for examples of usage. The happy path will probably
be moved to documentation tests when I'll be polishing API docs.