Replace PrivateKey/PublicKey with Vec<u8>
Using newtypes will cause friction when using SecureMessage. I liked
the idea of being generic over AsRef<[u8]> in Secure Cell. This gives
the user freedom to choose how to handle keys: they can either lend
them to Themis, or transfer ownership, or use shared keys, etc.

Newtypes have their merits: they deny the possibility of misusing
a private key in place of a public one. However, this should be rare
as people can be (sometimes) assumed to be reasonable. So we can be
more permissive.

Just in case, here's a test to make sure that Themis will complain
at least in run-time if the user misplaces their keys. Unfortunately,
Themis crashes with this test at the moment. So maybe we will get back
to using newtypes some day.