New Secure Message interface
This is the Rusty interface for Secure Messages. Currently sign-verify
mode may be a bit unwieldy as one needs to instantiate SecureSign and
SecureVerify structs to get access to their only methods. However, I
think that these contexts are meant to be reused so we'll provide them.
We can always add utility functions later.