Initial Secure Session wrapper
This is the first attempt at wrapping a Secure Session. It is somewhat
complicated by the callback interface. Currently this is dealt with a
single trait object cast into C's 'void*' type. However, maybe we'll
split it into multiple traits later: the only required one is the
get_public_key_by_id() routine, all other are optional.

The implementation is currently very crude and can be improved upon, but
I'll leave for later. It also supports only no-transport mode at the
moment. And there is only one big integration-scale test. And a bunch of
TODOs left unresolved. Well, I guess I'll leave it as is because that's
a week worth of job for finally undestanding how to wrap and use the
Secure Session.