Initial transport mode of Secure Session
This implements the transport mode of Secure Session, the one that uses
user-provided callbacks to send and receive data. After some testing I
have found multiple things which are wrong with this mode. So it's not
that surprising that most of other wrappers do not provide it :)

First of all, the callbacks feel pretty weird and it's not very useful
to have a maximum limit on the received message. Error handling is also
awful and will require updates in Themis to report errors correctly.
Finally, the connection sequence is not very user-friendly. It still
requires effectively calling negotiate() until the connection is
complete. Maybe we can hide it behind a neat interface later.

Anyway, here goes another couple of days of work or so. We'll improve
the interface and implementation a bit later.