Initial wrapper for Secure Comparator
This is initial alpha-quality wrapper for Secure Comparator. It has
some tests, some crude API, and it kinda works. There are multiple
things which can be improved here, mostly in the API. We'll get to
that later. I guess, when I try writing example code or somewhere
around that mark. For now I think I should focus on getting things
ready for the first release.

New tests also confirmed that Themis is somewhat unstable when faced
with corrupted data. I believe that at some point we should perform
some fuzz testing of Themis to weed out some of the issues.