Examples of Themis usage
This provides examples of Themis usage from Rust. Currently we cover
Keygen, Secure Cell, and Secure Message objects. Secure Session and
Secure Comparator are coming soon. (I'm wa-a-ay behind schedule at
the moment so I've decided to push whatever is ready.)

Writing these examples turned out to be valuable experience for me.
For one thing, I realized how astonishingly complicated the Tokio
framework is. I have wasted around two weekends trying to write
something simple for the chat example, but without any progress in
understanding the resulting code while making it short. As you can
see, I've given up and used stdlib for the example. It may not be
fast, but at least the examples are around 100 lines of code (as
opposed to 500 for my best attempt with Tokio).

(It's not like I'm giving up on learning Tokio. It is the current
industry standard after all. However, I won't be using this project
as a playground because it slows down my progress.)

Back to serious business.

- The examples have shown that the error types should exported
  from the root crate because of ergonomics.

- They have also hinted that Secure Message types could be made
  cloneable (though this is not used in the current versions
  of the examples).

- Finally, they made me realize the profound depth of the API
  choice for Secure Messages in wrap/unwrap vs. sign/verify mode.

  Keeping them different actually helps preventing misuse.
  Sign/verify does not encrypt the message content so switching
  from it to wrap/unwrap will require actual code changes rather
  than runtime configuration. This should make the developer think
  about the implications of using sign/verify vs. wrap/unwrap.