Initial Travis CI integration
Currently rust-themis requires Themis to be installed in the system
with default paths. Maybe in the future we'll make it easier for
Rust programs to integrate with it. However, maybe this will not
be good as we'll still need things like OpenSSL present.

We have to use sudo-enabled builds because Travis is crap and does
not allow installation on random packages via APT addon. It only
allows a specific whitelist (and has no resources for reviewing
additions into that whitelist). Thus we can't use container-based
infrastructure which kinda sucks (1 minute startup time for the
build instead of 5 seconds). Whatever...

This build has pretty basic matrix as we're just testing things out.
Later I'll add more stuff like lints, nightly builds, etc.