Initial API documentation
Finally add documentation comments where they belong. These are more or
less preliminary, they could definitely be improved, but I do expect API
changes so I don't really want to invest that much into comments now.

Note that we enable "missing_docs" warning to ensure that all public
items are documented. This is a warning (not an error) so that local
development is not really affected by it, but keep in mind that the CI
server treats warnings as errors and will fail the build.

Also note that some items are marked with "doc(hidden)" attribute. These
are mostly some internal error variants as well as some Drop trait impls
which are not important details for the userts.

Finally, we strive to document every public API (that is, everything
marked with "pub" in the source code), even if it is not actually
exported from the crate. Not only this is useful for development, but it
will come handy in case we will export these things later.