Themis 0.0.1
The first release of Themis for Rust. This is mostly so that we can
finally claim a name on crates.io, but the crate is actually usable.
All core API are available:

- Key generation
- Secure Cell
- Secure Message
- Secure Session
- Secure Comparator

This release is versioned 0.0.1, but it is actually compatible with
Themis 0.10.0. We'll use the 0.0.X versioning for a while, until the
API is more or less stabilized.

I dedicate this release to my beloved witch, Beatrice.