Add basic FFI bindings with bindgen
Up to this point we have been using manually specified C declarations.
This approach is not really robust, it will require manual updates when
new Themis core is released. This could be tedious and error-prone
(e.g., with constant values).

We're going to use more automated approach with using "bindgen" crate to
generate FFI bindings. The definitions are then included into our crate.