Move raw FFI bindings into a *-sys crate
This is Cargo convention of putting the wrappers into a separate crate
which enables better dependency management and isolation of concerns.

Usually the *-sys crate will also attempt building the wrapped library
if it is not present in the system, it is also expected to specify
native dependencies of the library in Cargo files as well. Currently
we do not do this but we probably should. Will do later.

The crate is imported as "bindings" to avoid changing usages all over
the code. But I may reconsider this as just spell out libthemis_sys.

We include some basic sanity tests as well as the license and readme
so that the crate looks nicer when published, but it is not expected
to be used on its own.