Fix clippy lints on the generated bindings
There's currently an outstanding issue with Clippy which applies lints
to path dependencies of a crate. This should not really happen. And it
breaks our build because bindgen generates some crap which make clippy
unhappy.

Disable the relevant lints in libthemis-sys crate. As noted in the
comment, we use the old form usable with stable which requires us
to manually spell out each lint we want disabled. On nightly we could
use #[allow(clippy::all)], but I do want to build on stable Rust.