Use panics for unrecoverable errors
Just panic if we encounter an error which cannot be reasonably prevented
by Themis users. For example, one cannot influence the memory allocator
directly so there is nothing that user code can do to prevent or handle
memory allocation failure.

This is somewhat big decision (for a library) because the user has way
to handle these errors. However, I do not want to litter program code
with useless error checks are almost never triggered and are most likely
to be simply unwrapped. If the need arises I have already added try_*
variants of the functions. These could be exported for the users if
someone actually needs to handle these kinds of failures.

Clippy insists on implementing Default when a type provides default-like
`fn new() -> Self` constructor. Do it to avoid warnings.