Rename private keys into secret keys
This was triggering me for quite a while and now I have found the
solution. By consistently using the word 'secret' we are now able
to observe perfectly aligned columns:

    let mut secret_key = Vec::new();
    let mut public_key = Vec::new();
    let mut secret_key_len = 0;
    let mut public_key_len = 0;

due to the irrefutable fact that words 'secret' and 'public' have
the same length. Additionally, secret and public are more visually
distinct than private and public (both starting with 'p').

This change should not affect the API. It affects documentation
but only because 'secret-private pair' reads dumb.