Use owned keys in Secure Cell and Message
Currently (Rust) Themis allows the user to pass as a key anything that
can be converted into a byte slice (i.e., T: AsRef<[u8]>). It means
that the user can choose to either give us a byte slice, or to pass
the whole ownership over the key. Which is kinda neat as it avoids
unnecessary copying.

I came up with this idea very early and stuck to it dearly, but it
has some disadvantages. Namely, you have to make everything generic
and lift the `T: AsRef<[u8]>` trait bound to data structures. This
will get even more annoying when trying to introduce a derivative
of T such as 'T which is an RSA key'. I am going to do precisely
that, and I tried really hard to preserve it but generics were way
more annoying than they should be.

Furthermore, if you look into the C code then you will see that the
key material gets copied around quite frequently so our optimizations
in Rust code do not have any measurable merit that would offset the
burden of dealing with generics everywhere.

Hence we get rid of these trait bounds. Now Secure Cell and Secure
Message always own the keys by making a copy of them. The keys are
stored as a new `KeyBytes` struct which acts as a marker for key
material. In the future we could for example make it securely zero
the memory when dropped so as to avoid leaking the keys.

Secure Session also uses keys but its story is a more complicated.
We'll get back to after we're done refactoring the keys to be more
strictly typed.

This change does not affect the API. All constructors of Secure Cell
and Secure Message still have the AsRef<[u8]> trait bound on their
arguments. However, their types are no longer generic and no longer
depend on the types of stuff passed into constructors, which relaxes
the lifetime mumbo-jumbo in case of constructing keys from slices.