Implement Debug trait for key types
It is a good idea to have Debug for *all* public types but I am too lazy
to audit the code for that right now. Debug pop out as a requirement in
some surprising places, like expect() method of Result, for example.
So add Debug for our key types, this will be useful in tests.

Note however that KeyBytes has a custom implementation of Debug. This
prevents leaks of key material in logs which are very likely to simply
use format!("{:?}", key).