Consistently use 'kind' for RSA vs ECDSA keys
There is a bit of a confusion in usage of words 'type' and 'kind'
w.r.t. keys. Initially I wanted to use the word 'type' to distinguish
between ECDSA and RSA keys. However, 'type' is a reserved keyword in
Rust so you can't use it for variable and method names. That's why
I've used 'kind' instead (as a related term from the type theory).
However, 'type' is still used to refer to actual static key types:
e.g., RsaPublicKey, KeyPair.

Let's consistently use words 'type' to refer to Rust types and 'kind'
to refer to values of the KeyKind enum.