Validating and constructing keys from bytes
This is the main dish. Here we teach rust-themis to construct strongly
typed keys from arbitrary byte slices. The slices are validated to
ensure that they indeed contain valid representations of Themis keys.

The most funny thing here is that Themis does not export any functions
to verify keys directly. I have tried (really) to write them in Rust,
but it turned out to be very frustrating. During those attempts I have
discovered that Themis does some slightly bad and unsafe things with
pointers which have discouraged me from even trying to pull that off
in Rust. It required too many unsafe's and #[allow]'s to get rustc and
clippy shut up about genuine undefined behavior and unsafe code.
You can read a more detailed rant in wrapper.c's comments.

So we're doing all those things in C now! This is not the best thing
to do because rust-themis now *requires* a C compiler to build. Before
this change we could be using a prebuilt libthemis which obviously does
not require a C compiler installed in the system. I hope that I'll
manage to move the C code into the core library at some point so that
we could drop the "cc" dependency.

While we're here, let's make KeyPair::try_join() more generic by doing
the {Rsa,Ecdsa}{Secret,Public}Key -> {Secret,Public}Key conversion
automatically. This allows the tests to be more concise and seems
to be a good API.

Yes, tests. Here are some tests to verify that keys are constructed
correctly and can handle malformed data. The tests have a few wacky
points which should away when keygen module is updated.