Use strict key types in Secure Session
Secure Session can work only with ECDSA keys. This was the main reason
for implementing strict key types. Now we are able to statically tell
the user to pass only ECDSA keys to Secure Session.

While we're here, make interface of SecureSessionTransport a bit
better by returning an optional key instead of using an out-parameter
with a mutable slice. This requires copying the key (as you may see
in tests), but it's more... Rusty. I hope we'll improve the API more
at some later point.