Introduce libthemis-src crate
We replicate the build model of `openssl` crate:

- `openssl` contains high-level Rust bindings. It wraps unsafe FFI
  functions provided by `openssl-sys` to expose a safe interface.

- `openssl-sys` provides raw FFI bindings. This crate is tasked
  with locating and linking to a native OpenSSL library, or building
  and embedding one if OpenSSL is not present on the system.

- `openssl-src` embeds the source code and is able to build it
  and install the artifacts to provided location.

`openssl` [depends] on `openssl-sys`.
`openssl-sys` [build-depends] on `openssl-src`.

Currently rust-themis is a separate repository so we embed Themis
source code a submodule. We target 0.10.0 release so use it.
Later I guess we'll have to have some Cargo trickery in order to
get it package the source code into published the crate. We'll
tackle this problem later.