Unnecessarily complex Themis location logic
Now we're going to have a bit more complex approach at locating the
Themis library, outputting more detailed instructions to Cargo. Yes,
we definitely could just ignore these issues and simply require the
users to install Themis somehow so that Cargo and company finds it.
But we're trying to be more helpful and nice. (And I personally want
to get some experience with complex build scripts. Here, I confessed.
Now let me do it.)

At the moment we only look in some standard UNIX paths for Themis.
Later commits will actually add more options. Stay tuned.

Currently we completely ignore Windows. I'm not even sure that Themis
can be built on Windows. So let's ignore it for a moment. It is true
suffering to do any software development on Windows.