Get location hints from Homebrew
Leverage Homebrew to locate Themis installation. Currently, Themis
does not provide official Homebrew formulas, but I personally do
and they are likely to be called "themis" or "themis-openssl" so
we use these names. We ask Homebrew to give us the installation
prefix (which usually is something like /usr/local/opt/themis)
and then probe the location for headers and libraries.

Note the HOMEBREW_NO_AUTO_UPDATE environment variable. This disables
automatic updates of Homebrew when running any "brew" commands.
The "--prefix" key should not trigger it but let's be safe.