Get location hints from pkg-config
This is mostly to support Linux systems that may reasonably rely on
pkg-config for locating libraries. It's not that used on macOS because
of Homebrew. Plus as far as I know official Linux packages for Themis
do not include *.pc files as well. However, pkg-config may be used so
let's support it (in some primitive form).

Note that we disable automatic printing of Cargo metadata. We're going
to double-check after pkg-config and decide on the library type
ourselves, so we do not need this support. We only want the prefix.