Check Themis version in libthemis-sys
We'd like to ensure that we link against the version we actually
support. Use themis_version() for that.

There are plans for removing this function from Themis. I guess we'll
have to abandon this check if the library does not provide any means
to verify the version. But currently the function is here so we can
use it.