Vendored Themis build
Support building our own statically-linked version of libthemis.
This is an option for the case when libthemis is not available
on the system and you don't want to install it.

The build depends on the `libthemis-src` crate for the source code.
It is an optional dependency enabled via "vendored" feature. The
feature is propagated to the core "themis" crate. You need to build
rust-themis with `--features "vendored"` in order to enable usage
of vendored source code.

Adding this dependency uncovered a bug in libthemis-src crate. Its
tests explicitly set the output directory but in libthemis-sys the
directory is taken from OUT_DIR which is then extended with "themis".
This subdirectory needs to be created before the build.