Use vendored BoringSSL in libthemis-src
Actually, do not let the user configure engine at all. If they
want a custom build of Themis they can do it themselves and tell
libthemis-sys to use it. Vendored build of rust-themis should
'just work' without any 3rd-party dependencies in the system.
That's why we should use the vendored BoringSSL version which
is used for Themis on Android.

Using Rust for build scripts is SO MUCH more verbose than shell.
Well, whatever... Also, BoringSSL uses CMake for its build
system. And Go.

It would probably be better to use `cmake` crate for invoking
CMake. However, I tried and it does not really work well with
the tests for `libthemis-src`. So maybe later.