Fix release vendored build of Themis
It turned out that in release builds Cargo sets a DEBUG environment
variable to "false". Themis' build system simply checks whether
DEBUG is defined, it does not look if it's set to "true" or "0".
Let's clear the variable in release builds to make sure they are
built without debugging enabled.