Drop vendored Themis build
After some careful consideration and considerable amounts of suffering
I have decided to remove the vendored build. It gives an impression of
being simple for the end user, but there are actually lots of details
that they must provide for the build to actually work. They must have
all compilers and stuff, CMake and Golang for BoringSSL. Plus it takes
around 2 minutes to build.  Furthermore, the build silently embeds
BoringSSL into the binary which can cause all sorts of mayhem due to
being binary-incompatible with OpenSSL while exporting the same symbol
names from the library.

So the vendored build has to go. Users of rust-themis will have to
provide the native library. This approach is used my some of other
Themis bindings and it should really be as easy as installing the
native package with your package manager.