Use pkg-config exclusively
I've been trying to get static libraries working for two weeks and I've
got enough of this crap. pkg-config is almost two decades old by now.
Please be a good UNIX citizen and provide *.pc files for your libraries.
If you're a special snowflake who has to build their own libraries then
go and write some *.pc files yourself and set PKG_CONFIG_PATH to them.
I don't want to maintain FIVE approaches for locating a single library
(with only one of them -- pkg-config -- actually supporting static
linkage in a good way).

So drop almost everything from the build script of libthemis-sys and
rely on pkg-config exclusively to find the library and print Cargo
instructions for us. See how few lines of code it takes? That's why.
We require "libthemis" and "libsoter" packages to be available.

Why does Soter need to be linked explicitly in when Themis should
depend on it and pull it in? That's because of "themis_shims" (see
wrapper.c) use symbols from libsoter so we have to link against it.
We can avoid this with static linkage everywhere and with dynamic
linkage on Linux due to way they work, but dynamic linkage on macOS
must be strict.

Also, don't use Homebrew for guessing the location of the libraries.
Normally it installs into /usr/local and the library will be available
by default. In case of keg-only installations (like OpenSSL) you have
to do special trickery anyway. Let the end user resolve these issues.
They should set PKG_CONFIG_PATH appropriately.