Document key generation
Well, this is easy. Grab the basis of documentation for the core Themis
wiki, provide some example of generating and using keys. Note when each
key type should be used.

We do not document the "keys" module here. It needs attention as well,
but I would like to focus on actually useful cryptosystem documentation
first.