Document Secure Cell
Improve the docs by reusing information from wiki and providing some
examples of usage, explaining caveats, etc.

Mind the TODO in context imprint documentation. Maybe it will make more
sense to panic, but I'm not sure. I have left this TODO as a reminder
to think about the API later. We keep (and document) the current
behavior when encryption fails if the provided context is empty.