Initial harness for generating rust-themis docs
Unfortunately, docs.rs can't generate and host Themis documentation
for us. This is because we require core Themis library to be installed
but it is not available on docs.rs. Therefore we have to build and host
our docs ourselves.

Here's a quick and dirty script which will download the source code
and build up-to-date documentation from it. It can pull releases from
crates.io as well as the latest development branch. I don't really
intend to publish master branch docs regularly, but it may come in
handy at times.

We use several hacks (also used by docs.rs) to add a 'header' with
version switch to the documentation. That's not optimal but it kinda
works.

The script writes it output to the current directory. It is intended
to be called manually from the checkout of gh-branches.

It also keeps a few "cache" directories: "_crates" and "_target" with
downloaded crates and Cargo output. They can be safely deleted and
will be regenerated, but keeping them saves a decent amount of time on
documentation updates.

Note the underscores in directory names. This ensures that GitHub Pages
do not host these files on the web site.

This commit also includes several other 'technical' files:

- .gitignore should be self-explanatory
- CNAME file to point GitHub Pages to our domain
- images/* with some static logos
- index.html with a user-side redirect to the latest version
  of the docs. It's nice to have so that you can type
  "https://rust-themis.ilammy.net/" and not get an HTTP 404.

We are going to host the docs on rust-themis.ilammy.net (a GitHub Pages
site on my domain). For the time being, at least. Later it should be
moved to a more nicely looking domain.

This commit does not include any automatically generated files. These
files are intended to be edited manually. Later commits will contain
actual docs.