Restore vendored build
It turns out that vendored builds may be useful for docs.rs so we bring
it back.

Unfortunately, Themis is not available in standard Linux repositories
and is not preinstalled on most of the systems. Thus docs.rs is unable
to build documentation for Themis because "cargo doc" requires
compilation of dependent crates.

I have tried 'solving' this issue by building documentation manually
(see gh-pages branch in this repo, specifically commit
75dccfe9a329693f3b0497f6be8451274e67b759). However, this won't scale and
I don't really like how this way ties the docs to my domain. Ideally the
API documentation should be hosted by either upstream (Cossack Labs) or
by docs.rs.

Vendored build will allow building Themis on docs.rs and thus should
enable documenting the root "themis" crate. Therefore we revert commit
497ffccb165e0cf0b6f2dbc460211b5ba7f6dcdd and improve it.

As before, the source code is split into a dedicated crate
"libthemis-src" (mimicking the "openssl" crate). It is used by
"libthemis-sys" if the "vendored" feature is enabled. Note that
currently the vendored build does not replace the system library. We
will still prefer the system library if it is available. Also note that
we always link statically against the vendored library.

We are not going to locate the library manually. We stick with using
pkg-config for that. Themis version pinned via submodule is the latest
master which does install *.pc files correctly. (Currently I use my own
fork which fixes another issue with ldconfig. Later the submodule should
be updated to the upstream master.)

Travis build script has been updated to check the vendored build without
core Themis library installed.

Note that I do not expect the "vendored" feature to be used for anything
other than docs.rs. Yes, this is a hack. Yes, it may even work. However,
transitive dependencies are hard to enfore with vendored builds so I'm
not seriously considering this as a supported option.