Update to Rust 2018
Rust 1.31 (and Rust 2018) have been stabilized and this has some impact
on the tool behavior.

- Update formatting to the new rustfmt standard

- Use scoped clippy attributes and properly ignore all warnings
  in bindgen-generated code.

- Stabilized clippy brings in some new warnings. Squelch a couple
  of them in our code base where they are not appropriate.

After that mark our crates as requiring Rust 2018 edition and update
the code accordingly. This mostly amounts to fixing up crate-local
imports to "use crate::*"

Note our update to Rust 2018 in the changelog. Strictly speaking, it
is not a breaking change as Rust editions do not change the API/ABI.
However, Themis now requires at least rustc 1.31 with Rust 2018 support
to compile. Though, it's not that we have announced any commitment
to support older rustc versions (we support the latest stable, just
as the core Rust team).