Securely zero key material before freeing memory
It is a good security practice to wipe sensitive data from memory as
soon as it no longer needed. Obviously, the data has to be in memory
for some time but we should not keep it around for more than absolutely
necessary. The data in a memory region is not erased when the memory
is freed so we have to make sure it's gone ourselves.

Initially I was overconfident and tried writing my own zeroing. Don't
do this, kids. There's a whole load of issues that need to be resolved
in order to *securely* zero out a memory region. You can read the
implementation details in the source code of "zeroize" or elsewhere.

Therefore we use a dedicated library for the task. It makes sure that
the compiler does not optimize away the code that tries its best to
write zeros into the main memory (not the caches). Currently zeroize
does not use these fancy functions like memset_s(), explicit_bzero(),
RtlSecureZeroMemory(), etc. but it might it the future.

There's also a test to verify the the memory is indeed zeroed out.
By definition, the test contains some technically unsafe code, but
it's unlikely to crash or misbehave.