Drop embedded pkg-config files
The latest Themis source code includes the changes which make embedded
pkg-config files obsolete. Drop the *.pc files, update hte README.

The files were using during the CI build via PKG_CONFIG_PATH variable.
Drop this variable as the files are no longer available. Unfortunately,
this breaks "System Themis" builds which still use the latest published
version 0.10. Therefore allow this build to fail (temporarily).
Introduce a new "Master Themis" build in order to compensate for the
loss. This build will use the latest Themis from the core repo.

The new build uses 'install' section to install Themis. This overrides
the global install section which was used to install various Rust
components. Move them to a more appropriate 'before_install' section.

While we're here, update the vendored version of Themis and retarget it
to the core repository. Updating submodules is a bit strange (which is
the reason why I do not like them), so you'd better just do a clean git
clone of the repo as "git submodule sync && git submodule update" may
not work as expected when remote URL changes.

Updates in the yet-to-be-0.11 Themis change the behavior of reusing
Secure Comparator so update the tests accordingly. This will fail with
Themis 0.10, but it's okay because rust-themis was never released and
will support only Themis 0.11+. Later we will have to take more care
about backwards compatibility, but this time we can let it slip.